Add range check for 'N' parameter
Colin Percival has a 'sanity check' that 0 < logN < 256: see
https://github.com/Tarsnap/scrypt/blob/484dc1fb/lib/scryptenc/scryptenc.c#L206.

Since the 'N' parameter in node-scrypt is the canonical scrypt logN, it is
useful to trap the canonical 'N' being passed as early as possible, and to
report it more helpfully than 'error computing derived key'.

Following this commit, if a call is erroneously made to
    scrypt.kdf('abc', { N: 65536, r: 8, p: 1 })
it will immediately have a RangeError thrown.